Move `NakamotoBlockProposal` into postblock_proposal.rs
